# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TR4CE Evidence API",
    "version": "1.0.0",
    "description": "Vault evidence derived from confirmed on-chain observations. Every report cites the exact rows it rests on, and a repeated request over identical observations returns the report that already exists rather than a new one."
  },
  "paths": {
    "/v1/vaults": {
      "get": {
        "operationId": "listVaults",
        "summary": "List curated vaults.",
        "parameters": [
          {
            "name": "chainId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 8453
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The curated registry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultListResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request body or query does not match the schema.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/reports": {
      "post": {
        "operationId": "createReport",
        "summary": "Produce an evidence report, or return the one that already exists.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "An equivalent report was already stored; nothing was created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            }
          },
          "201": {
            "description": "A new report was stored.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request body or query does not match the schema.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "No such vault or report.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "The observations or chain state needed to answer do not allow it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/reports/{id}": {
      "get": {
        "operationId": "getReport",
        "summary": "Read a stored report exactly as it was served.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^trc_[0-9a-f]{32}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The stored report.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            }
          },
          "404": {
            "description": "No such vault or report.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/policies/evaluate": {
      "post": {
        "operationId": "evaluatePolicy",
        "summary": "Evaluate a candidate policy without storing anything.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluatePolicyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The evaluation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluatePolicyResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request body or query does not match the schema.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "No such vault or report.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "The observations or chain state needed to answer do not allow it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "The candidate policy did not validate; issues are returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluatePolicyResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/actions/prepare": {
      "post": {
        "operationId": "prepareAction",
        "summary": "Build the unsigned calls for a deposit or redemption, and simulate the first.",
        "description": "Returns unsigned transactions only. TR4CE never signs and never submits: the wallet owner approves each call, and only the first unsent call is simulated because the second reverts until the first has landed.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrepareActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The prepared action, with the first call simulated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreparedActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request body or query does not match the schema.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "No such vault or report.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "The observations or chain state needed to answer do not allow it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "501": {
            "description": "This deployment has no chain configured for actions.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/actions/{id}": {
      "get": {
        "operationId": "getActionStatus",
        "summary": "Whether a prepared action may still be signed, judged now.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^act_[0-9a-f]{32}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The action's standing.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionStatusResponse"
                }
              }
            }
          },
          "404": {
            "description": "No such vault or report.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "501": {
            "description": "This deployment has no chain configured for actions.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/actions/{id}/submitted": {
      "post": {
        "operationId": "reportActionSubmission",
        "summary": "Report the transaction hash a wallet produced.",
        "description": "The only route by which a transaction hash enters TR4CE, and it arrives as a report about something that already happened in the caller's wallet.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^act_[0-9a-f]{32}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportSubmissionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The action's standing after the report.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionStatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request body or query does not match the schema.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "No such vault or report.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "The observations or chain state needed to answer do not allow it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "501": {
            "description": "This deployment has no chain configured for actions.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiError": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "INVALID_REQUEST",
                  "UNKNOWN_VAULT",
                  "INSUFFICIENT_OBSERVATIONS",
                  "INCOMPLETE_EVIDENCE",
                  "INVALID_POLICY",
                  "REPORT_NOT_FOUND",
                  "ACTION_NOT_AVAILABLE",
                  "ACTION_NOT_FOUND",
                  "INTERNAL_ERROR"
                ]
              },
              "message": {
                "type": "string",
                "minLength": 1
              },
              "reasonCodes": {
                "type": "array",
                "items": {
                  "type": "string",
                  "enum": [
                    "MISSING_OBSERVATION",
                    "STALE_EVIDENCE",
                    "INCOMPATIBLE_ASSET",
                    "INCOMPATIBLE_IMPLEMENTATION",
                    "CALL_REVERTED",
                    "UNSUPPORTED_CAPABILITY",
                    "AMBIGUOUS_CAPABILITY",
                    "UNSUPPORTED_VAULT",
                    "PROVIDER_UNAVAILABLE",
                    "REORG_INVALIDATED",
                    "INVALID_POLICY",
                    "SIMULATION_REVERTED",
                    "SIMULATION_STALE",
                    "WALLET_CONTEXT_CHANGED"
                  ]
                }
              },
              "issues": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "required": [
                    "path",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            },
            "required": [
              "code",
              "message",
              "reasonCodes",
              "issues"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "VaultListResponse": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "schemaVersion": {
            "type": "string",
            "const": "1.0.0"
          },
          "vaults": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "chainId": {
                  "type": "integer",
                  "exclusiveMinimum": 0,
                  "maximum": 9007199254740991
                },
                "address": {
                  "type": "string",
                  "pattern": "^0x[0-9a-fA-F]{40}$"
                },
                "symbol": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "name": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "asset": {
                  "type": "string",
                  "pattern": "^0x[0-9a-fA-F]{40}$"
                },
                "assetSymbol": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "shareDecimals": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 9007199254740991
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "candidate",
                    "listed",
                    "degraded",
                    "unsupported"
                  ]
                },
                "adapterKey": {
                  "type": "string"
                },
                "adapterVersion": {
                  "type": "string"
                }
              },
              "required": [
                "chainId",
                "address",
                "symbol",
                "name",
                "asset",
                "assetSymbol",
                "shareDecimals",
                "status",
                "adapterKey",
                "adapterVersion"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "schemaVersion",
          "vaults"
        ],
        "additionalProperties": false
      },
      "CreateReportRequest": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "chainId": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          },
          "vaultAddress": {
            "type": "string",
            "pattern": "^0x[0-9a-fA-F]{40}$"
          },
          "windowDays": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 365
          },
          "policy": {
            "type": "object",
            "properties": {
              "version": {
                "type": "number",
                "const": 1
              },
              "underlyingAssets": {
                "minItems": 1,
                "type": "array",
                "items": {
                  "type": "string",
                  "const": "USDC"
                }
              },
              "minHistoryDays": {
                "type": "integer",
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991
              },
              "minTvlAssets": {
                "type": "string",
                "pattern": "^(0|[1-9][0-9]*)$"
              },
              "minObservedReturnBps": {
                "type": "object",
                "properties": {
                  "windowDays": {
                    "type": "integer",
                    "exclusiveMinimum": 0,
                    "maximum": 9007199254740991
                  },
                  "value": {
                    "type": "integer",
                    "minimum": -9007199254740991,
                    "maximum": 9007199254740991
                  }
                },
                "required": [
                  "windowDays",
                  "value"
                ],
                "additionalProperties": false
              },
              "minWithdrawableAssets": {
                "type": "object",
                "properties": {
                  "owner": {
                    "type": "string",
                    "pattern": "^0x[0-9a-fA-F]{40}$"
                  },
                  "value": {
                    "type": "string",
                    "pattern": "^(0|[1-9][0-9]*)$"
                  }
                },
                "required": [
                  "owner",
                  "value"
                ],
                "additionalProperties": false
              }
            },
            "required": [
              "version",
              "underlyingAssets",
              "minHistoryDays",
              "minTvlAssets",
              "minObservedReturnBps",
              "minWithdrawableAssets"
            ],
            "additionalProperties": false
          },
          "policyName": {
            "default": "default",
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          }
        },
        "required": [
          "chainId",
          "vaultAddress",
          "windowDays",
          "policy",
          "policyName"
        ],
        "additionalProperties": false
      },
...
      "EvaluatePolicyRequest": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "chainId": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          },
          "vaultAddress": {
            "type": "string",
            "pattern": "^0x[0-9a-fA-F]{40}$"
          },
          "windowDays": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 365
          },
          "policy": {}
        },
        "required": [
          "chainId",
          "vaultAddress",
          "windowDays",
          "policy"
        ],
        "additionalProperties": false
      },
      "EvaluatePolicyResponse": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "schemaVersion": {
            "type": "string",
            "const": "1.0.0"
          },
          "evaluation": {
            "anyOf": [
              {
                "type": "object",
                "properties": {
                  "version": {
                    "type": "number",
                    "const": 1
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "PASS",
                      "FAIL",
                      "UNKNOWN"
                    ]
                  },
                  "rules": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "key": {
                          "type": "string",
                          "enum": [
                            "underlyingAsset",
                            "minimumHistory",
                            "minimumTvl",
                            "minimumObservedReturn",
                            "minimumWithdrawableAssets"
                          ]
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "PASS",
                            "FAIL",
                            "UNKNOWN"
                          ]
                        },
                        "threshold": {
                          "type": "string"
                        },
                        "observedValue": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "evidenceReferences": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "reasonCodes": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "enum": [
                              "MISSING_OBSERVATION",
                              "STALE_EVIDENCE",
                              "INCOMPATIBLE_ASSET",
                              "INCOMPATIBLE_IMPLEMENTATION",
                              "CALL_REVERTED",
                              "UNSUPPORTED_CAPABILITY",
                              "AMBIGUOUS_CAPABILITY",
                              "UNSUPPORTED_VAULT",
                              "PROVIDER_UNAVAILABLE",
                              "REORG_INVALIDATED",
                              "INVALID_POLICY",
                              "SIMULATION_REVERTED",
                              "SIMULATION_STALE",
                              "WALLET_CONTEXT_CHANGED"
                            ]
                          }
                        }
                      },
                      "required": [
                        "key",
                        "status",
                        "threshold",
                        "observedValue",
                        "evidenceReferences",
                        "reasonCodes"
                      ],
                      "additionalProperties": false
                    }
                  }
                },
                "required": [
                  "version",
                  "status",
                  "rules"
                ],
                "additionalProperties": false
              },
              {
                "type": "null"
              }
            ]
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "path",
                "message"
              ],
              "additionalProperties": false
            }
          },
          "asOfBlock": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(0|[1-9][0-9]*)$"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "schemaVersion",
          "evaluation",
          "issues",
          "asOfBlock"
        ],
        "additionalProperties": false
      },
...
      "ReportSubmissionRequest": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "chainId": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          },
          "transactionHash": {
            "type": "string",
            "pattern": "^0x[0-9a-fA-F]{64}$"
          }
        },
        "required": [
          "chainId",
          "transactionHash"
        ],
        "additionalProperties": false
      }
    }
  }
}